+++ server.ts
@@ -1,6 +1,17 @@
-import server from './src'
-const port = process.env.PORT || 3000
+import PonyGen from './src'
 
-server.listen(port, () => {
-  console.log(`Example app listening at http://localhost:${port}`)
+let port:number
+
+if (process.env.PORT) {
+  port = Number.parseInt(process.env.PORT, 10);
+} else {
+  port = 3000
+}
+
+const server = new PonyGen()
+
+server.listen(port).then(() => {
+  console.log(`Server started on http://localhost:${port}`)
+}).catch((err) => {
+  console.error(`Error starting server: ${err}`)
 })