+++ src/index.spec.ts
@@ -1,8 +1,45 @@
 import request from "supertest";
-import server from "./";
+import PonyGen from "./";
 
-it("Index", async (done) => {
-  request(server)
+it("Server - can start and stop", async (done) => {
+  const server = new PonyGen();
+  expect(server.server()).toBeNull
+  expect(await server.listen(2001)).resolves
+  expect(server.isRunning()).toBeTruthy;
+
+  await server
+    .close()
+    .then(() => {
+      expect(server.isRunning()).toBeFalsy;
+      done();
+    })
+    .catch((err) => {
+      expect(err).toBeFalsy;
+    });
+
+});
+
+it("Server - can start", async (done) => {
+  const server = new PonyGen();
+  server
+    .listen(1818)
+    .then(() => {
+      server.close();
+      done();
+    })
+    .catch((err) => {
+      expect(err).toBeFalsy;
+    });
+});
+
+it("Server - can throw on bad close", async (done) => {
+  const server = new PonyGen();
+  await expect(server.close()).rejects.toThrowError(/not running/);
+  done();
+});
+
+it("Index", (done) => {
+  request(new PonyGen()._express)
     .get("/about")
     .expect(200)
     .end((err, res) => {
@@ -13,7 +50,7 @@ it("Index", async (done) => {
 });
 
 it("API", async (done) => {
-  request(server)
+  request(new PonyGen()._express)
     .get("/api")
     .expect(200)
     .end((err, res) => {
@@ -24,7 +61,7 @@ it("API", async (done) => {
 });
 
 it("Raw API - home", async (done) => {
-  request(server)
+  request(new PonyGen()._express)
     .get("/api/raw")
     .expect(200)
     .end((err, res) => {
@@ -35,7 +72,7 @@ it("Raw API - home", async (done) => {
 });
 
 it("Raw API - pony", async (done) => {
-  request(server)
+  request(new PonyGen()._express)
     .get("/api/raw/pony/1")
     .expect("Content-Type", /json/)
     .expect(200)