+++ src/index.ts
@@ -1,10 +1,36 @@
-import express from 'express'
-import routeHome from './routes'
+import express from "express";
+import { Server } from "node:http";
+import routeHome from "./routes";
 
-const app = express()
+export default class PonyGen {
+  _isRunning = false;
+  _express = express();
+  _server: null | Server = null;
 
-app.use('/', routeHome)
+  constructor() {
+    this._express.use("/", routeHome);
 
-app.use(express.static('public'))
+    this._express.use(express.static("public"));
+  }
 
-export default app
+  async listen(port: number): Promise<void> {
+    this._server = await this._express.listen(port)
+    this._isRunning = true;
+  }
+
+  async close(): Promise<void> {
+    if (this._server === null) {
+      throw new Error("Attempted to close server when it was not running");
+    }
+    await this._server.close()
+    this._isRunning = false;
+  }
+
+  isRunning(): boolean {
+    return this._isRunning;
+  }
+
+  server(): null | Server {
+    return this._server;
+  }
+}